refactor(libs-shared-infra-fetch-server): rename server files to remove .server suffix
- Rename batch-fetch.server.ts to batch-fetch.ts
- Rename fetch-with-auth.server.ts to fetch-with-auth.ts
- Update import paths in dependent files
