+++ libs/shared/infra/fetch-server/src/index.ts
@@ -1,3 +1,3 @@
-export { resetBatchFetcher } from './batch-fetch.server'
-export * from './fetch-with-auth.server'
+export { resetBatchFetcher } from './batch-fetch'
+export * from './fetch-with-auth'
 export { gqlServerRequest } from './gql-request.server'
